requirements.txt
emms-sdk>=0.27.0
deepagents>=0.4.0
